Working camera and daemon containers
