+++ camera/common/udev.py
